Add `quickSort` as a new sorting algorithm
Add a Quicksort implementation, it works but I now noticed that I dont
have actual tests that I can see anyhow. It would be good to have some
unit test and run these in a CI from once in a while to make sure these
are doing good when I come and check them. Also, the build was broken
and I didnt notice.